// lib/tdex/types.ts
export type Coin = Asset & {
  amount?: number
}

export interface CoinPair {
  from: Coin
  dest: Coin
}

export interface TDEXv2Market {
  provider: TDEXv2Provider
  baseAsset: string
  baseAmount?: string
  quoteAsset: string
  quoteAmount?: string
  percentageFee?: { baseAsset: string; quoteAsset: string }
  fixedFee?: { baseAsset: string; quoteAsset: string }
  price?: TDEXv2MarketPrice
}

/**
 * Checks if given market is a TDEXv2Market
 * @param market
 * @returns market is TDEXv2Market
 */
export function isTDEXv2Market(market: any): market is TDEXv2Market {
  return (
    typeof market === 'object' &&
    typeof market.baseAsset === 'string' &&
    typeof market.quoteAsset === 'string'
  )
}

export interface TDEXv2Provider {
  name: string
  endpoint: string
}

/**
 * Checks if given provider is a TDEXv2Provider
 * @param provider
 * @returns provider is TDEXv2Provider
 */
export function isTDEXv2Provider(provider: any): provider is TDEXv2Provider {
  return (
    typeof provider === 'object' &&
    typeof provider.name === 'string' &&
    typeof provider.endpoint === 'string'
  )
}

export interface TDEXv2MarketPrice {
  spotPrice: number
  minTradableAmount: string
}